Исправил замечания.

+++ src/main/java/com/fitness/courses/http/coach/course/model/info/CourseEntityWithStudentsCount.java
@@ -2,6 +2,6 @@ package com.fitness.courses.http.coach.course.model.info;
 
 import com.fitness.courses.http.coach.course.model.entity.CourseEntity;
 
-public record CourseEntityWithStudentsCount(CourseEntity course, Integer studentsCount)
+public record CourseEntityWithStudentsCount(CourseEntity course, Long studentsCount)
 {
 }
